update docs/readme/scripts to match new grammar
