+++ scripts/coins.xpl
@@ -1,11 +1,11 @@
 * coins.xpl
 * summation of money values
 
-@|?h {p n d q .}
-+|+13 {hey guys!}
-h@|+ {H}            * halt =
-r#|3 {p} {x}    * rule =
-r#|3 {n} {xxxxx}      * rule =
-r#|3 {d} {xxxxxxxxxx}       * rule =
-r#|3 {q} {xxxxxxxxxxxxxxxxxxxxxxxxx}
-n?#+ {p.p.n.d.H}          * input =
+@|?h p n d q .
++|+13 hey guys!
+h@|+ H            
+r#|3 p x   
+r#|3 n x x x x x       
+r#|3 d x x x x x x x x x x      
+r#|3 q x x x x x x x x x x x x x x x x x x x x x x x x x
+n?#+ p . p . n . d . H
