+++ scripts/infinite.xpl
@@ -1,7 +1,7 @@
 * infinite.xpl
 * infinitely looping program
 
-@|?h {a}         * alph =
-h@|+ {H}         * halt =
-r#|3 {a} {aa}    * rule =
-n?#+ {aaa}       * input =
+@|?h a
+h@|+ H
+r#|3 a a a
+n?#+ a a a
