+++ scripts/test.xpl
@@ -1,9 +1,16 @@
 * Test.xpl
 * Sample from tag system wiki
 
-@|?h {a b c d} 		* alph =
-h@|+ {H}            * halt =
-r#|3 {a} {ccbaH}    * rule =
-r#|3 {b} {cca}      * rule =
-r#|3 {c} {cc}       * rule =
-n?#+ {baa}          * input =
+@|?h a b c d	
+h@|+ H         
+r#|3 a c c b a H
+r#|3 b c c a 
+r#|3 c c c 
+
+* to match wiki page, use -i baa when running
+* to set the initial tape configutation to baa
+* Note: when using command-line tape configs,
+* do not add spaces between characters!
+* therefore, command line configs will only
+* work if you have defined the symbols in the
+* alphabet to be only 1 character long
